Build a common components module on webpack build
Will look through each folder in the app/components/ folder, and generate a app/components/components.js file that exports
a angular module with all found components already imported.

This logic is run on webpack build. But does not yet watch for changes to the components. So if you add a component
you have to restart your build.

If you use the 1508cli tool to add the component, you need to go to the console anyway, so that is not a big
deal

#14