Add watcher to generated content
Script that is started in parallel with the webpack build to watch for new folders in  `components`, `filters` and `services` - and then generated
the corresponding angular module.

This replaces the previous webpack shell plugin powered script to generate on webpack builds. That solution did not pick up new folders, and you
would have to stop and restart the webpack dev server to get new things added.